feat: create hook for dragscroll, change slider to height
