+++ src/pages/ui-components/Tabs/index.tsx
@@ -9,6 +9,7 @@ import IconButton from '../IconButton';
 import nestedComponent from '../utils/nestedComponent';
 import { ChevronLeft, ChevronRight } from './Icons';
 import Tab from './Tab';
+import useTabDragScroller from './useTabDragScroller';
 import useTabSteps from './useTabSteps';
 import Vertical from './Vertical';
 
@@ -38,6 +39,8 @@ const Tabs = ({ children, value, onChange, selectOnMobile, ...rest }: TabsProps)
 
   const labelsRef = React.useRef<HTMLDivElement>(null);
   const parentRef = React.useRef<HTMLDivElement>(null);
+
+  const { touchStartHandler } = useTabDragScroller(parentRef?.current as HTMLElement);
 
   const isMobile = useMediaQuery(breakpoints.down('sm'));
 
@@ -111,34 +114,6 @@ const Tabs = ({ children, value, onChange, selectOnMobile, ...rest }: TabsProps)
     },
     [onChange]
   );
-
-  const ele = parentRef?.current as HTMLElement;
-
-  const [scrolledPosition, setScrolledPosition] = React.useState({ scrollLeft: 0, x: 0 });
-
-  const touchStartHandler = (e: React.TouchEvent) => {
-    const position = {
-      scrollLeft: ele.scrollLeft,
-      x: e.touches[0].clientX
-    };
-
-    setScrolledPosition(position);
-
-    window.addEventListener('touchmove', touchMoveHandler as any);
-    window.addEventListener('touchend', touchEndHandler);
-  };
-
-  const touchMoveHandler = (e: React.TouchEvent<HTMLDivElement>) => {
-    console.log(e, 'eventHandler');
-    const { scrollLeft, x } = scrolledPosition;
-    const howMuchMoved = e.touches[0].clientX - x;
-    ele.scrollLeft = scrollLeft - howMuchMoved;
-  };
-
-  const touchEndHandler = () => {
-    window.removeEventListener('touchmove', touchMoveHandler as any);
-    window.removeEventListener('touchend', touchEndHandler);
-  };
 
   if (isMobile && selectOnMobile) {
     return (
@@ -241,9 +216,8 @@ const TabsWrapper = React.memo(
       .hst-tabs__slider {
         position: absolute;
         transition: all 0.2s;
-        border: solid;
-        border-color: ${theme.brandColor.primary.pure};
-        border-width: ${theme.border.width.xs};
+        height: ${theme.border.width.sm};
+        background-color: ${theme.brandColor.primary.pure};
       }
 
       .hst-tabs__tab {
